adds more tests
