+++ ee/src/test/java/io/supertokens/ee/test/EETest.java
@@ -1,15 +1,16 @@
 package io.supertokens.ee.test;
 
-import com.google.gson.JsonObject;
 import io.supertokens.ProcessState;
-import io.supertokens.ee.test.httpRequest.HttpRequestForTesting;
 import io.supertokens.featureflag.EE_FEATURES;
 import io.supertokens.featureflag.FeatureFlag;
-import io.supertokens.webserver.WebserverAPI;
 import org.junit.*;
 import org.junit.rules.TestRule;
 
 public class EETest {
+
+    public static final String OPAQUE_LICENSE_KEY = "TODO";
+    public static final String OPAQUE_INVALID_LICENSE_KEY = "abcd";
+
 
     @Rule
     public TestRule watchman = Utils.getOnFailure();
@@ -25,43 +26,7 @@ public class EETest {
     }
 
     @Test
-    public void testRetrievingEEFeaturesWhenNotSetInDb() throws Exception {
-        String[] args = {"../../"};
-
-        TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
-        Assert.assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
-
-        // check that there are no features enabled
-        EE_FEATURES[] eeArray = FeatureFlag.getInstance(process.getProcess()).getEnabledFeatures();
-        Assert.assertEquals(0, eeArray.length);
-
-        // check that isLicenseKeyPresent is false
-        Assert.assertFalse(
-                FeatureFlag.getInstance(process.getProcess()).getEeFeatureFlagInstance().getIsLicenseKeyPresent());
-        process.kill();
-        Assert.assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
-    }
-
-    @Test
-    public void testThatCallingGetFeatureFlagAPIReturnsEmptyArray() throws Exception {
-        String[] args = {"../../"};
-
-        TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
-        Assert.assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
-
-        JsonObject response = HttpRequestForTesting.sendGETRequest(process.getProcess(), "",
-                "http://localhost:3567/ee/featureflag",
-                null, 1000, 1000, null, WebserverAPI.getLatestCDIVersion(), "");
-        Assert.assertEquals("OK", response.get("status").getAsString());
-        Assert.assertNotNull(response.get("features"));
-        Assert.assertEquals(0, response.get("features").getAsJsonArray().size());
-
-        process.kill();
-        Assert.assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
-    }
-
-    @Test
-    public void testRemovingLicenseKey() throws Exception {
+    public void testRemovingLicenseKeyWhenItIsNotSet() throws Exception {
         String[] args = {"../../"};
 
         TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
@@ -84,35 +49,42 @@ public class EETest {
     }
 
     @Test
-    public void testDeletingLicenseKeyWhenItDoesNotExist() throws Exception {
+    public void testNoNetworkCallMadeWhenNoLicenseKey() throws Exception {
         String[] args = {"../../"};
 
         TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
         Assert.assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
 
-        JsonObject response = HttpRequestForTesting.sendJsonDELETERequest(process.getProcess(), "",
-                "http://localhost:3567/ee/license",
-                null, 1000, 1000, null, WebserverAPI.getLatestCDIVersion(), "");
-        Assert.assertEquals("OK", response.get("status").getAsString());
+        Assert.assertNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.LICENSE_KEY_CHECK_NETWORK_CALL, 2000));
 
         process.kill();
         Assert.assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
     }
 
     @Test
-    public void testRetrievingLicenseKeyWhenItIsNotSet() throws Exception {
+    public void testNetworkCallMadeWhenLicenseKeyPresent() throws Exception {
         String[] args = {"../../"};
 
-        TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
-        Assert.assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+        {
+            TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
+            Assert.assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
 
-        JsonObject response = HttpRequestForTesting.sendGETRequest(process.getProcess(), "",
-                "http://localhost:3567/ee/license",
-                null, 1000, 1000, null, WebserverAPI.getLatestCDIVersion(), "");
-        Assert.assertEquals("NO_LICENSE_KEY_FOUND_ERROR", response.get("status").getAsString());
+            FeatureFlag.getInstance(process.main).setLicenseKeyAndSyncFeatures(OPAQUE_LICENSE_KEY);
 
-        process.kill();
-        Assert.assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+            process.kill();
+            Assert.assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+        }
+
+        {
+            TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
+            Assert.assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+
+            Assert.assertNotNull(
+                    process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.LICENSE_KEY_CHECK_NETWORK_CALL));
+
+            process.kill();
+            Assert.assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+        }
     }
 }
