+++ ee/src/test/java/io/supertokens/ee/test/LicenseKeyAPITest.java
@@ -47,4 +47,36 @@ public class LicenseKeyAPITest {
         process.kill();
         Assert.assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
     }
+
+    @Test
+    public void testRetrievingLicenseKeyWhenItIsNotSet() throws Exception {
+        String[] args = {"../../"};
+
+        TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
+        Assert.assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+
+        JsonObject response = HttpRequestForTesting.sendGETRequest(process.getProcess(), "",
+                "http://localhost:3567/ee/license",
+                null, 1000, 1000, null, WebserverAPI.getLatestCDIVersion(), "");
+        Assert.assertEquals("NO_LICENSE_KEY_FOUND_ERROR", response.get("status").getAsString());
+
+        process.kill();
+        Assert.assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+    }
+
+    @Test
+    public void testDeletingLicenseKeyWhenItDoesNotExist() throws Exception {
+        String[] args = {"../../"};
+
+        TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
+        Assert.assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+
+        JsonObject response = HttpRequestForTesting.sendJsonDELETERequest(process.getProcess(), "",
+                "http://localhost:3567/ee/license",
+                null, 1000, 1000, null, WebserverAPI.getLatestCDIVersion(), "");
+        Assert.assertEquals("OK", response.get("status").getAsString());
+
+        process.kill();
+        Assert.assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+    }
 }
